tools/requirements.txt
pyftdi>=0.55.0
pyspiflash>=0.6.3
